Migrated to Core Data with some bugs

+++ Expenses/View Controllers/InfoViewController.swift
@@ -40,7 +40,7 @@ class InfoViewController: UIViewController, UIDocumentMenuDelegate, UIDocumentPi
                     let columns = row.components(separatedBy: "\t")
                     if columns.count >= 6 {
                         let date = dateFormat.date(from: columns[0])
-                        let amount = (columns[1] as NSString).floatValue
+                        let amount = (columns[1] as NSString).doubleValue
                         let accountName = columns[2]
                         let category = columns[3]
                         let project = columns[4]
